requirements.txt
python-telegram-bot
astrapy
python-dotenv
flask
flask-socketio
psutil
requests
aiohttp
